backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
bcrypt==4.1.1
pyjwt==2.8.1
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httpx==0.25.2
python-jose==3.3.0
pytz==2023.3
cryptography==41.0.7
